+++ Changelog.md
@@ -1,3 +1,6 @@
+## 0.0.12
++ Minor stability updates.
+
 ## 0.0.11
 + Added support for Vercel Edge runtime.
 + Fixed `Parameter.enumeration` field type to be `EnumerationMember[]` instead of `EnumerationMember`.